LayoutTests/fast/speech/input-ontextinput-event-expected.txt

Tests for the ontextInput event with <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("speechInput").value is ""
PASS window.event.data is "Green eggs and ham"
PASS window.event.inputMethod is undefined
PASS successfullyParsed is true

TEST COMPLETE
